pyproject.toml
[tool.poetry]
name = "fsa-test"
version = "0.1.0"
description = ""
authors = ["Jacob Shore <jshore@datalign.net>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.3"
python-dotenv = "^0.13.0"
flask-bcrypt = "^0.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
